Django_API/drive_api/requirements.txt
debugpy==1.6.5
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.1.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.90.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
importlib-metadata==5.0.0
Markdown==3.4.1
mssql-django==1.3
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.0
pandas==1.5.1
pickleshare==0.7.5
PyDrive==1.3.1
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
SQLAlchemy==2.0.17
sqlparse==0.4.3
